{{cookiecutter.project_name}}/requirements.txt
cdktf~=0.16.1
psycopg2-binary==2.9.6
sshtunnel==0.4.0
pytest
